Fix Dropdown Editor Focus Bug
* Fixed WinForms control stealing focus from dropdowns in World Editor.
* Overworld tunnels now properly use the cliff palette.
